requirements.txt
openai
exa-py
python-dotenv
